+++ rest/api/v2010/model_api_v2010_incoming_phone_number_mobile.go
@@ -3,7 +3,7 @@
  *
  * This is the public Twilio REST API.
  *
- * API version: 1.20.3
+ * API version: 1.21.0
  * Contact: support@twilio.com
  */
 
@@ -32,7 +32,7 @@ type ApiV2010IncomingPhoneNumberMobile struct {
 	DateUpdated *string `json:"date_updated,omitempty"`
 	// The emergency address configuration to use for emergency calling
 	EmergencyAddressSid *string `json:"emergency_address_sid,omitempty"`
-	// Whether the phone number is enabled for emergency calling
+	// Displays if emergency calling is enabled for this number.
 	EmergencyStatus *string `json:"emergency_status,omitempty"`
 	// The string that you assigned to describe the resource
 	FriendlyName *string `json:"friendly_name,omitempty"`