+++ rest/taskrouter/v1/workspaces_tasks_reservations.go
@@ -3,7 +3,7 @@
  *
  * This is the public Twilio REST API.
  *
- * API version: 1.20.3
+ * API version: 1.21.0
  * Contact: support@twilio.com
  */
 
@@ -192,6 +192,8 @@ func (c *ApiService) getNextListTaskReservationResponse(nextPageUrl string) (int
 
 // Optional parameters for the method 'UpdateTaskReservation'
 type UpdateTaskReservationParams struct {
+	// The If-Match HTTP request header
+	IfMatch *string `json:"If-Match,omitempty"`
 	// Whether to play a notification beep when the participant joins or when to play a beep. Can be: `true`, `false`, `onEnter`, or `onExit`. The default value is `true`.
 	Beep *string `json:"Beep,omitempty"`
 	// Whether to play a notification beep when the customer joins.
@@ -300,6 +302,10 @@ type UpdateTaskReservationParams struct {
 	WorkerActivitySid *string `json:"WorkerActivitySid,omitempty"`
 }
 
+func (params *UpdateTaskReservationParams) SetIfMatch(IfMatch string) *UpdateTaskReservationParams {
+	params.IfMatch = &IfMatch
+	return params
+}
 func (params *UpdateTaskReservationParams) SetBeep(Beep string) *UpdateTaskReservationParams {
 	params.Beep = &Beep
 	return params
@@ -686,6 +692,10 @@ func (c *ApiService) UpdateTaskReservation(WorkspaceSid string, TaskSid string, 
 	}
 	if params != nil && params.WorkerActivitySid != nil {
 		data.Set("WorkerActivitySid", *params.WorkerActivitySid)
+	}
+
+	if params != nil && params.IfMatch != nil {
+		headers["If-Match"] = *params.IfMatch
 	}
 
 	resp, err := c.requestHandler.Post(c.baseURL+path, data, headers)